# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/catalogs": {
            "get": {
                "tags": [
                    "Catalogs"
                ],
                "summary": "Retrieve all catalogs",
                "operationId": "42de0014c181b2db4e8be53300e91171",
                "responses": {
                    "200": {
                        "description": "List of all catalogs"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Retrieve all categories",
                "operationId": "14ac1e1b5090300ecf261902e78e0b58",
                "responses": {
                    "200": {
                        "description": "List of all categories"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/sub-categories": {
            "get": {
                "tags": [
                    "Sub-Categories"
                ],
                "summary": "Retrieve all sub-categories",
                "operationId": "38e785daace0abd75e866ace7981d53a",
                "responses": {
                    "200": {
                        "description": "List of all sub-categories"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/top-products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Retrieve top products",
                "operationId": "7e2a5eb2bd97381c691a0c1bce0d298f",
                "responses": {
                    "200": {
                        "description": "List of all top products"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Catalogs",
            "description": "Catalogs"
        },
        {
            "name": "Categories",
            "description": "Categories"
        },
        {
            "name": "Sub-Categories",
            "description": "Sub-Categories"
        },
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}